requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
google-generativeai==0.3.1
chromadb==0.4.18
python-dotenv==1.0.0
